+++ include/mockturtle/algorithms/resubstitution.hpp
@@ -42,6 +42,7 @@
 #include "detail/mffc_utils.hpp"
 #include "reconv_cut.hpp"
 #include "simulation.hpp"
+#include "reconv_cut2.hpp"
 
 #include <fmt/format.h>
 #include <kitty/dynamic_truth_table.hpp>
@@ -55,160 +56,6 @@ namespace mockturtle
 
 namespace detail
 {
-  /*** reconvergence-driven cut based on abcReconv.c ***/
-  template<typename Ntk>
-  struct cut_manager
-  {
-    cut_manager( int node_size_max, int node_fan_stop = 100000 )
-      : node_size_max( node_size_max )
-      , node_fan_stop( node_fan_stop )
-    {
-    }
-
-    /* \brief limit on the size of the supernode */
-    int node_size_max;
-
-    /* \brief limit on the size of the containing cone */
-    int cone_size_max;
-
-    /* \brief limit on the size of the supernode */
-    int node_fan_stop;
-
-    /* \brief fanins of the collapsed node (the cut) */
-    std::vector<node<Ntk>> node_leaves;
-
-    /* \brief fanins of the containing cone */
-    std::vector<node<Ntk>> cone_leaves;
-
-    /* \brief visited nodes */
-    std::vector<node<Ntk>> visited;
-
-    /* \brief data structure to compute TFO nodes */
-    std::vector<std::vector<node<Ntk>>> levels;
-
-    /* \brief nodes in the TFO of the cut */
-    std::vector<node<Ntk>> nodes_tfo;
-  };
-
-  template<typename Ntk>
-  int node_get_leaf_cost_one( Ntk const& ntk, typename Ntk::node const &node, int fanin_limit )
-  {
-    /* make sure the node is in the construction zone */
-    assert( ntk.visited( node ) == 1 );
-
-    /* cannot expand over the PI node */
-    if ( ntk.is_constant( node ) || ntk.is_pi( node ) ) // TODO: is_ci
-      return 999;
-
-    /* get the cost of the cone */
-    uint32_t cost = 0;
-    ntk.foreach_fanin( node, [&]( const auto& f ){
-        cost += ntk.visited( ntk.get_node( f ) ) ? 0 : 1;
-      } );
-
-    /* always accept if the number of leaves does not increase */
-    if ( cost < ntk.fanin_size( node ) )
-      return cost;
-
-    /* skip nodes with many fanouts */
-    if ( int( ntk.fanout_size( node ) ) > fanin_limit )
-      return 999;
-
-    /* return the number of nodes that will be on the leaves if this node is removed */
-    return cost;
-  }
-
-  template<typename Ntk>
-  bool node_build_cut_level_one_int( Ntk const& ntk, std::vector<typename Ntk::node>& visited, std::vector<typename Ntk::node>& leaves, uint64_t size_limit, int fanin_limit )
-  {
-    uint32_t best_cost = 100;
-
-    std::optional<typename Ntk::node> best_fanin;
-    int best_pos;
-
-    /* evaluate fanins of the cut */
-    auto pos = 0;
-    for ( const auto& l : leaves )
-    {
-      uint32_t cost_curr = node_get_leaf_cost_one( ntk, l, fanin_limit );
-      if ( best_cost > cost_curr ||
-           ( best_cost == cost_curr && best_fanin && ntk.level( l ) > ntk.level( *best_fanin ) ) )
-      {
-        best_cost = cost_curr;
-        best_fanin = std::make_optional( l );
-        best_pos = pos;
-      }
-
-      if ( best_cost == 0 )
-        break;
-
-      ++pos;
-    }
-
-    if ( !best_fanin )
-      return false;
-
-    // assert( best_cost < max_fanin_of_graph_structure );
-    if ( leaves.size() - 1 + best_cost > size_limit )
-        return false;
-
-    /* remove the best node from the array */
-    leaves.erase( leaves.begin() + best_pos );
-
-    /* add the fanins of best to leaves and visited */
-    ntk.foreach_fanin( *best_fanin, [&]( const auto& f ){
-        auto const& n = ntk.get_node( f );
-        if ( !ntk.visited( n ) )
-        {
-          ntk.set_visited( n, 1u );
-          visited.push_back( n );
-          leaves.push_back( n );
-        }
-      });
-
-    assert( leaves.size() <= size_limit );
-
-    return true;
-  }
-
-  template<typename Ntk>
-  void node_unmark( Ntk const& ntk, std::vector<typename Ntk::node>& visited )
-  {
-    for ( const auto& v : visited )
-    {
-      ntk.set_visited( v, 0u );
-    }
-  }
-
-  template<class Ntk>
-  std::vector<typename Ntk::node> node_find_cut( cut_manager<Ntk>& mgr, Ntk const& ntk, typename Ntk::node const& root )
-  {
-    /* start the visited nodes and mark them */
-    mgr.visited.clear();
-    mgr.visited.push_back( root );
-    ntk.set_visited( root, 1 );
-    ntk.foreach_fanin( root, [&]( const auto& f ){
-        auto const& n = ntk.get_node( f );
-        mgr.visited.push_back( n );
-        ntk.set_visited( n, 1 );
-      } );
-
-    /* start the cut */
-    mgr.node_leaves.clear();
-    ntk.foreach_fanin( root, [&]( const auto& f ){
-        auto const& n = ntk.get_node( f );
-        mgr.node_leaves.push_back( n );
-      } );
-
-    /* compute the cut */
-    while ( node_build_cut_level_one_int( ntk, mgr.visited, mgr.node_leaves, mgr.node_size_max, mgr.node_fan_stop ) );
-    assert( int( mgr.node_leaves.size() ) <= mgr.node_size_max );
-
-    /* unmark fMarkB in tbe TFI */
-    node_unmark( ntk, mgr.visited );
-    return mgr.node_leaves;
-  }
-
   /*** observability don't cares based on abcOdc.c ***/
   struct odc_parameters
   {
